changes to project due to tests

+++ PropertyBuildingDemo.Api/Controllers/GenericEntityController.cs
@@ -12,6 +12,7 @@ using Microsoft.EntityFrameworkCore;
 using PropertyBuildingDemo.Domain.Common;
 using PropertyBuildingDemo.Domain.Entities.Enums;
 using PropertyBuildingDemo.Domain.Interfaces;
+using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
 
 /// <summary>
 /// Base controller for handling CRUD operations for generic entities of type TEntity.
@@ -39,26 +40,32 @@ public class GenericEntityController<TEntity, TEntityDto> : BaseController
     }
 
     /// <summary>
-    /// Retrieves a list of entities of type TEntity.
+    /// Gets a list of entities with optional filtering for deleted entities.
     /// </summary>
-    /// <returns>An IActionResult containing the list of entities or an error message.</returns>
+    /// <param name="filterDeleted">A boolean indicating whether to include deleted entities in the list.</param>
+    /// <returns>A list of entities.</returns>
     [HttpGet("List")]
-    public async Task<IActionResult> Index()
+    public async Task<IActionResult> Index([FromQuery] bool filterDeleted = false)
     {
         ApiResult<IEnumerable<TEntityDto>> apiResult = null;
         try
         {
             // Retrieve all entities of type TEntity from the database
-            var dataList = await this._unitOfWork.GetRepository<TEntity>().Entities.ToListAsync();
+            var dataList = this._unitOfWork.GetRepository<TEntity>().Entities;
+            if (!filterDeleted)
+            {
+                // If filterDeleted is false, exclude deleted entities.
+                dataList = dataList.Where(x => !x.IsDeleted);
+            }
 
-            apiResult = ApiResult<IEnumerable<TEntityDto>>.SuccessResult(
-                this._mapper.Map<IEnumerable<TEntityDto>>(dataList)
-                );
+            apiResult = await ApiResult<IEnumerable<TEntityDto>>.SuccessResultAsync(
+                this._mapper.Map<IEnumerable<TEntityDto>>(await dataList.ToListAsync())
+            );
         }
         catch (Exception ex)
         {
             // Handle exceptions and log them
-            apiResult = ApiResult<IEnumerable<TEntityDto>>.FailedResult(ex.Message);
+            apiResult = await ApiResult<IEnumerable<TEntityDto>>.FailedResultAsync(ex.Message);
             this._systemLogger.LogExceptionMessage(ELoggingLevel.Error, $"{nameof(TEntity)}/Index failed!", ex);
         }
 
@@ -70,7 +77,7 @@ public class GenericEntityController<TEntity, TEntityDto> : BaseController
     /// </summary>
     /// <param name="id">The ID of the entity to retrieve.</param>
     /// <returns>An IActionResult containing the entity or an error message if not found.</returns>
-    [HttpGet("ById")]
+    [HttpGet("ById/{id}")]
     public async Task<IActionResult> GetById(long id)
     {
         ApiResult<TEntityDto> apiResult = null;
@@ -79,14 +86,19 @@ public class GenericEntityController<TEntity, TEntityDto> : BaseController
             // Retrieve the entity by its ID using the unit of work and repository
             var result = await this._unitOfWork.GetRepository<TEntity>().GetAsync(id);
 
+            if (result.IsDeleted)
+            {
+                result = null;
+            }
+
             // Check if the entity was found or not, and create an appropriate API result
-            apiResult = result == null ? ApiResult<TEntityDto>.FailedResult() : 
-                ApiResult<TEntityDto>.SuccessResult(this._mapper.Map<TEntityDto>(result));
+            apiResult = result == null ? await ApiResult<TEntityDto>.SuccessResultAsync("not found") :
+                await ApiResult<TEntityDto>.SuccessResultAsync(this._mapper.Map<TEntityDto>(result));
         }
         catch (Exception ex)
         {
             // Handle exceptions and log them
-            apiResult = ApiResult<TEntityDto>.FailedResult(ex.Message);
+            apiResult = await ApiResult<TEntityDto>.FailedResultAsync(ex.Message);
             this._systemLogger.LogExceptionMessage(ELoggingLevel.Error, $"{nameof(TEntity)}/GetById failed!", ex);
         }
 
@@ -114,13 +126,13 @@ public class GenericEntityController<TEntity, TEntityDto> : BaseController
             bool succeeded = await this._unitOfWork.Complete();
 
             // Create an appropriate API result based on the operation's success
-            apiResult = result == null ? ApiResult<TEntityDto>.FailedResult($"Failed to insert {nameof(TEntityDto)}") :
-                ApiResult<TEntityDto>.SuccessResult(this._mapper.Map<TEntityDto>(result));
+            apiResult = result == null && succeeded ? await ApiResult<TEntityDto>.FailedResultAsync($"Failed to insert {nameof(TEntityDto)}") :
+                await ApiResult<TEntityDto>.SuccessResultAsync(this._mapper.Map<TEntityDto>(result));
         }
         catch (Exception ex)
         {
             // Handle exceptions and log them
-            apiResult = ApiResult<TEntityDto>.FailedResult(ex.Message);
+            apiResult = await ApiResult<TEntityDto>.FailedResultAsync(ex.Message);
             this._systemLogger.LogExceptionMessage(ELoggingLevel.Error, $"{nameof(TEntity)}/Insert failed!", ex);
         }
 
@@ -148,13 +160,13 @@ public class GenericEntityController<TEntity, TEntityDto> : BaseController
             bool succeeded = await this._unitOfWork.Complete();
 
             // Create an appropriate API result based on the operation's success
-            apiResult = result == null ? ApiResult<TEntityDto>.FailedResult($"Failed to Update {nameof(TEntity)}") :
-                ApiResult<TEntityDto>.SuccessResult(this._mapper.Map<TEntityDto>(result));
+            apiResult = result == null && succeeded ? await ApiResult<TEntityDto>.FailedResultAsync($"Failed to Update {nameof(TEntity)}") :
+                await ApiResult<TEntityDto>.SuccessResultAsync(this._mapper.Map<TEntityDto>(result));
         }
         catch (Exception ex)
         {
             // Handle exceptions and log them
-            apiResult = ApiResult<TEntityDto>.FailedResult(ex.Message);
+            apiResult = await ApiResult<TEntityDto>.FailedResultAsync(ex.Message);
             this._systemLogger.LogExceptionMessage(ELoggingLevel.Error, $"{nameof(TEntity)}/Update failed!", ex);
         }
 
@@ -166,31 +178,31 @@ public class GenericEntityController<TEntity, TEntityDto> : BaseController
     /// </summary>
     /// <param name="id">The ID of the entity to delete.</param>
     /// <returns>An IActionResult indicating success or failure.</returns>
-    [HttpDelete("Delete")]
+    [HttpDelete("Delete/{id}")]
 
-    public async Task<IActionResult> Delete([FromBody] long id)
+    public async Task<IActionResult> Delete( long id)
     {
         ApiResult<TEntityDto> apiResult = null;
         try
         {
             // Retrieve the entity by its ID
             var item = await this._unitOfWork.GetRepository<TEntity>().GetAsync(id);
-            if (item != null)
+            if (item != null && !item.IsDeleted)
             {
                 // Delete the entity and complete the unit of work transaction
                 await this._unitOfWork.GetRepository<TEntity>().DeleteAsync(item);
                 bool succeeded = await this._unitOfWork.Complete();
-                apiResult = !succeeded ? ApiResult<TEntityDto>.FailedResult($"Failed to Delete {nameof(TEntityDto)}") :
-                    ApiResult<TEntityDto>.SuccessResult(this._mapper.Map<TEntityDto>(item));
+                apiResult = !succeeded ? await ApiResult<TEntityDto>.FailedResultAsync($"Failed to Delete {nameof(TEntityDto)}") :
+                    await ApiResult<TEntityDto>.SuccessResultAsync(this._mapper.Map<TEntityDto>(item));
             }
             else
             {
-                apiResult = ApiResult<TEntityDto>.FailedResult($"Failed to Delete {nameof(TEntity)} because it does not exist");
+                apiResult = await ApiResult<TEntityDto>.FailedResultAsync($"Failed to Delete {nameof(TEntity)} because it does not exist");
             }
         }
         catch (Exception ex)
         {
-            apiResult = ApiResult<TEntityDto>.FailedResult(ex.Message);
+            apiResult = await ApiResult<TEntityDto>.FailedResultAsync(ex.Message);
             this._systemLogger.LogExceptionMessage(ELoggingLevel.Error, $"{nameof(TEntity)}/Update failed!", ex);
         }
